requirements.txt
fastapi
uvicorn
requests
dnspython
fpdf
httpx
pydantic